Updating third party dependencies and node version
Iniital version build only on NodeJS version v12. I sequentially
introduced some changes and now it builds on v16.
To do that I need to update some of the libraries
and do some tweaks in code. Some of changes like the ones in
ObjectParser.ts are just naive type casting.
In future I need to review that code, understand and fix proper way.