Removed ThemeWatcher
We don't need to track Themes in real time. In fact right now
we don't even need to handle themes at all. I left the code
for theme changing but for now no theme is loaded from directory
and default one (statically) is used in app. In future we
may support theme changing but now it is not crucial.