Change playlistWatcher to one time initialization on start
Ok, chokidar was a bad idea. It is not maintained. Current
FolderWatcher looks robust enough to keep with it. But I want
to simplify logic as much as possible. We really doesn't need
real live refresh of playlists. It will only happen when adding
new modules and this is good enough to restart exogui to refresh
playlists.