Fix for showing duplicates of extras
Extras were added to the list by two functions after being added
to the addApps list in the initial run.

I changed flow of funciton. I've added set which indicates if the
extras for the game were loaded. If not the initializaiton function
is called for the game loading all extras and adding to addApp list.

In that manner we are sure that we read only once the files.