Implement alternative image path with uid
When there are two games with same name the uid may be added to the
image filename.

Interesting bug with game constructor. To make it easier to match
image filenmaes with game name I lowercased and converted name.
And apparently constructor name used as a object index is not good
idea :D as it refers to object constructor. Changing tolowercase
for touppercase solved the error.